# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: GigFolio API — Developer 3 Scope
  description: |
    API specifications for Developer 3 deliverables in the GigFolio platform:
    - **Reputation Engine:** Breakdown and composite scoring
    - **Consent Management & Share Links:** Time-boxed, scope-limited access delegation
    - **QR Code Generation:** Authoritative backend generation for portable profile sharing
    - **Enterprise Verification:** Scoped data projection and SD-JWT presentation
    - **Application Audit Events:** Tamper-resistant, hashed-identifier access logs
  version: 1.0.0
  contact:
    name: GigFolio Engineering Team
    email: dev@gigfolio.io
servers:
  - url: https://api.gigfolio.io
    description: Production API
  - url: http://localhost:3000
    description: Local Development Server
paths:
...
  /api/v1/consent/share-links:
    post:
      tags:
        - Consent & Share Links
      summary: Create a share link with QR code
      description: |
        Creates a time-boxed, scope-limited share link. The returned `share_token`
        is an opaque raw token shown **EXACTLY ONCE**. The backend only persists
        its SHA-256 hash. TTL is strictly capped at 720 hours.
      operationId: createShareLink
      security:
        - BearerAuth:
            - worker:consent:write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShareLinkRequest'
            example:
              permitted_scopes:
                - identity:name_only
                - earnings:monthly_aggregate
                - reputation:composite_score
              ttl_hours: 72
              max_uses: 1
              passcode: optional-string
      responses:
        '201':
          description: Share link created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareLinkResponse'
              example:
                share_token: opaque-raw-token-shown-once
                share_url: https://gigfolio.io/verify/opaque-raw-token-shown-once
                qr_code_svg_url: https://gigfolio.io/qr/opaque-raw-token-shown-once.svg
                permitted_scopes:
                  - identity:name_only
                  - earnings:monthly_aggregate
                  - reputation:composite_score
                expires_at: '2026-09-06T12:00:00Z'
        '400':
          description: Invalid scope or TTL > 720 hours
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Worker account unverified or insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/consent/share-links/{id}:
    delete:
      tags:
        - Consent & Share Links
      summary: Revoke a share link
      description: |
        Revokes a share link immediately. Revocation is propagated to Redis
        so that subsequent verification attempts are rejected instantly.
      operationId: revokeShareLink
      security:
        - BearerAuth:
            - worker:consent:delete
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the share link to revoke
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Share link revoked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeShareLinkResponse'
              example:
                message: Share link revoked successfully.
                revoked_at: '2026-09-03T18:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Share link not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
  /api/v1/public/workers/{worker_public_id}:
    get:
      tags:
        - Public Worker Lookup
      summary: Lookup worker by public ID (pointer resolution)
      description: |
        Resolves a `worker_public_id` to the minimum discoverable profile.
        **This is a POINTER, not a bearer credential.**

        Only returns non-sensitive, worker-controlled fields. `worker_public_id` alone grants
        ZERO data access. To access actual data (reputation, earnings, identity), obtain a
        share token from the worker and call `POST /api/v1/organizations/verify-worker`.
      operationId: getPublicWorkerProfile
      security:
        - OrgApiKeyAuth: []
      parameters:
        - name: worker_public_id
          in: path
          required: true
          description: The stable public UUID the worker shared (NOT their internal DB UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Public profile resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicWorkerProfileResponse'
              example:
                worker_public_id: 7f3a2b1c-4e5d-6f7a-8b9c-0d1e2f3a4b5c
                display_name: Jane W.
                is_identity_verified: true
                has_reputation_score: true
                verification_badge: PLATFORM_VERIFIED
                data_access_hint: >-
                  To access actual worker data, obtain a share token from this worker
                  and call POST /api/v1/organizations/verify-worker.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No GigFolio profile found for this public ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/v1/workers/me/rotate-public-id:
    post:
      tags:
        - Worker Settings
      summary: Rotate worker public ID
      description: |
        Generates a new `worker_public_id` and immediately invalidates the old one.
        Use when the public ID has been shared with untrusted parties.
        Internal records (reputation, consent, audit history) are completely unaffected.
      operationId: rotateWorkerPublicId
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Public ID rotated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RotatePublicIdResponse'
              example:
                new_worker_public_id: a1b2c3d4-e5f6-7890-abcd-ef1234567890
                previous_public_id: 7f3a2b1c-4e5d-6f7a-8b9c-0d1e2f3a4b5c
                rotated_at: '2026-09-03T19:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Worker JWT containing worker ID and granted scopes.
    OrgApiKeyAuth:
      type: apiKey
      in: header
      name: X-Organization-Key
      description: Institutional Verifier API Key.
  schemas:
...
    CreateShareLinkRequest:
      type: object
      required:
        - permitted_scopes
        - ttl_hours
      properties:
        permitted_scopes:
          type: array
          items:
            type: string
          minItems: 1
          example:
            - identity:name_only
            - earnings:monthly_aggregate
            - reputation:composite_score
        ttl_hours:
          type: integer
          minimum: 1
          maximum: 720
          example: 72
        max_uses:
          type: integer
          nullable: true
          minimum: 1
          example: 1
        passcode:
          type: string
          example: optional-string
    ShareLinkResponse:
      type: object
      required:
        - share_token
        - share_url
        - qr_code_svg_url
        - permitted_scopes
        - expires_at
      properties:
        share_token:
          type: string
          description: Opaque raw token returned exactly once
          example: opaque-raw-token-shown-once
        share_url:
          type: string
          format: uri
          example: https://gigfolio.io/verify/opaque-raw-token-shown-once
        qr_code_svg_url:
          type: string
          format: uri
          example: https://gigfolio.io/qr/opaque-raw-token-shown-once.svg
        permitted_scopes:
          type: array
          items:
            type: string
        expires_at:
          type: string
          format: date-time
          example: '2026-09-06T12:00:00Z'
    RevokeShareLinkResponse:
      type: object
      required:
        - message
        - revoked_at
      properties:
        message:
          type: string
          example: Share link revoked successfully.
        revoked_at:
          type: string
          format: date-time
          example: '2026-09-03T18:00:00Z'
...
    ErrorResponse:
      type: object
      required:
        - statusCode
        - message
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
          nullable: true
    PublicWorkerProfileResponse:
      type: object
      required:
        - worker_public_id
        - has_reputation_score
        - data_access_hint
      properties:
        worker_public_id:
          type: string
          format: uuid
          description: The stable public handle (NOT the internal worker UUID)
          example: 7f3a2b1c-4e5d-6f7a-8b9c-0d1e2f3a4b5c
        display_name:
          type: string
          nullable: true
          example: Jane W.
        is_identity_verified:
          type: boolean
          nullable: true
          example: true
        has_reputation_score:
          type: boolean
          description: Whether a reputation score exists. Does NOT reveal the score.
          example: true
        verification_badge:
          type: string
          enum: [VERIFIED, PLATFORM_VERIFIED, UNVERIFIED]
          nullable: true
          example: PLATFORM_VERIFIED
        profile_updated_at:
          type: string
          format: date-time
          nullable: true
          example: '2026-09-03T12:00:00Z'
        data_access_hint:
          type: string
          description: >-
            Instructs the caller how to access actual data via the verify-worker flow.
    RotatePublicIdResponse:
      type: object
      required:
        - new_worker_public_id
        - previous_public_id
        - rotated_at
      properties:
        new_worker_public_id:
          type: string
          format: uuid
          description: The newly generated public ID. Share this going forward.
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        previous_public_id:
          type: string
          format: uuid
          description: The old public ID, now permanently invalidated.
          example: 7f3a2b1c-4e5d-6f7a-8b9c-0d1e2f3a4b5c
        rotated_at:
          type: string
          format: date-time
          example: '2026-09-03T19:00:00Z'
  responses:
    Unauthorized:
      description: Missing or invalid authentication credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Access denied or insufficient scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'